Form merge (design wise: form shouldn't change drastically anytime soon)
