+++ Foosball2text/Foosball2text.csproj
@@ -72,13 +72,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="VideoLoggerForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="VideoLoggerForm.Designer.cs">
-      <DependentUpon>VideoLoggerForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="VideoLoggerMessageDelivery.cs" />
+    <Compile Include="LoggerMessageDelivery.cs" />
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
@@ -91,9 +85,6 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="VideoLoggerForm.resx">
-      <DependentUpon>VideoLoggerForm.cs</DependentUpon>
-    </EmbeddedResource>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
