+++ Foosball2text/Form1.cs
@@ -6,6 +6,8 @@ using Emgu.CV.CvEnum;
 using Emgu.CV.Structure;
 using System.Drawing;
 using System.IO;
+using System.ComponentModel;
+using System.Linq;
 
 namespace Foosball2text
 {
@@ -15,23 +17,26 @@ namespace Foosball2text
         private const int _fps = 30;
         private VideoCapture _capture;
         private FrameHandler _frameHandler;
-        private VideoLoggerForm _logger;
         string _filePath;    
 
         public Form1()
         {
             InitializeComponent();
-            _logger = new VideoLoggerForm();
-            _logger.Show();
             _frameHandler = new FrameHandler(pictureBox1.Width, pictureBox1.Height);
             _timer = new Timer();
+
             //Frame Rate
             _timer.Interval = 1000 / _fps;
             _timer.Tick += new EventHandler(TimerTick);
             _timer.Start();
 
-            _filePath = "../../sample.avi"; //<---- On launch video file path
+            _filePath = "../../sample.avi"; //<---- On-launch-video file path
             _capture = new VideoCapture(_filePath);
+
+            //--- Logger
+            logData.Add(messageGetter.gameStart);
+            listBox1.DataSource = logData;
+            logData.ListChanged += new ListChangedEventHandler(list_ListChanged);
         }
 
         private void TimerTick(object sender, EventArgs e)
@@ -44,10 +49,10 @@ namespace Foosball2text
             imageBox1.Image = _frameHandler.GetFilteredImage(frame, pictureBox1.Width, pictureBox1.Height);
             UpdateCordinates();
 
-            //BallWatcher Update
+            //TODO rework BallWatcher
             _frameHandler.ballWatcher.UpdateBallWatcher();
             Teams teamScored = _frameHandler.ballWatcher.checkWhichTeamScored();
-            _logger.UpdateBallWatcherData(_frameHandler.ballWatcher.GetBallOnSideString(),
+            UpdateBallWatcherData(_frameHandler.ballWatcher.GetBallOnSideString(),
                                           _frameHandler.ballWatcher.GetSpeedString(),
                                           teamScored);
         }
@@ -56,7 +61,6 @@ namespace Foosball2text
         {
             _xlabel.Text = _frameHandler.X;
             _ylabel.Text = _frameHandler.Y;
-            _logger.UpdateBallCoordinates(_frameHandler._ball.X, _frameHandler._ball.Y);
         }
 
         private void pictureBox1_Click(object sender, EventArgs e)
@@ -91,14 +95,76 @@ namespace Foosball2text
                 try //to assign new _capture
                 {
                     _capture = new VideoCapture(_filePath);
-                    _logger.newGame();
+                    newGame();
                 }
                 catch (IOException ex)
                 {
                     MessageBox.Show(ex.ToString());
                 }
             }
-         }
+        }
+
+        // ************ Logger methods ************
+        BindingList<String> logData = new BindingList<string>();
+        LoggerMessageDelivery messageGetter = new LoggerMessageDelivery();
+
+        private void list_ListChanged(object sender, ListChangedEventArgs e)
+        {
+            LatestLogUpdate();
+        }
+
+        private void LatestLogUpdate()
+        {
+            LatestLog.Text = logData.Last();
+
+            int visibleItems = listBox1.ClientSize.Height / listBox1.ItemHeight;
+            listBox1.TopIndex = Math.Max(listBox1.Items.Count - visibleItems + 1, 0);
+        }
+
+        private void EndGameButton_Click(object sender, EventArgs e)
+        {
+            logData.Add(messageGetter.gameEnd);
+            ResetScore();
+        }
+
+        public void UpdateBallWatcherData(string ballOnSideText, string speedText, Teams teamScored)
+        {
+            ballOnSideOfFieldValue.Text = ballOnSideText;
+            SpeedValue.Text = speedText;
+
+            if (teamScored == Teams.TeamA)
+                AddGoalA();
+            if (teamScored == Teams.TeamB)
+                AddGoalB();
+        }
+
+        public void AddGoalA()
+        {
+            int score = int.Parse(TeamA.Text);
+            score++;
+            TeamA.Text = score.ToString();
+            logData.Add(messageGetter.goalLeft);
+        }
+
+        public void AddGoalB()
+        {
+            int score = int.Parse(TeamB.Text);
+            score++;
+            TeamB.Text = score.ToString();
+            logData.Add(messageGetter.goalRight);
+        }
+
+        public void ResetScore()
+        {
+            TeamA.Text = "0";
+            TeamB.Text = "0";
+        }
+
+        public void newGame()
+        {
+            ResetScore();
+            logData.Add(messageGetter.gameStart);
+        }
     }
 }
